create voter and anonymus user and roles works

+++ src/Controller/TaskController.php
@@ -9,6 +9,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\Exception\AccessDeniedException;
 use Symfony\Component\Security\Core\Security;
 
 class TaskController extends AbstractController
@@ -34,22 +35,26 @@ class TaskController extends AbstractController
     #[Route('/tasks/create', name: 'task_create')]
     public function create(Request $request, Security $security, EntityManagerInterface $em)
     {
-        $task = new Task();
-        $form = $this->createForm(TaskType::class, $task);
-        $form->handleRequest($request);
+        if ($this->getUser() != null) {
+            $task = new Task();
+            $form = $this->createForm(TaskType::class, $task);
+            $form->handleRequest($request);
 
-        if ($form->isSubmitted() && $form->isValid()) {
+            if ($form->isSubmitted() && $form->isValid()) {
+                $task->setAuthor($security->getUser());
+                $em->persist($task);
+                $em->flush();
 
-            $task->setAuthor($security->getUser());
-            $em->persist($task);
-            $em->flush();
+                $this->addFlash('success', 'La tâche a été bien été ajoutée.');
 
-            $this->addFlash('success', 'La tâche a été bien été ajoutée.');
-
-            return $this->redirectToRoute('task_list');
+                return $this->redirectToRoute('task_list');
+            }
+            return $this->render('task/create.html.twig', ['form' => $form->createView()]);
         }
 
-        return $this->render('task/create.html.twig', ['form' => $form->createView()]);
+        $this->addFlash('error', 'Vous devez être connecté pour créer une tâche.');
+
+        return $this->redirectToRoute('login');
     }
 
     /**
@@ -62,7 +67,6 @@ class TaskController extends AbstractController
     public function edit(Task $task, Request $request, EntityManagerInterface $em)
     {
         $form = $this->createForm(TaskType::class, $task);
-
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
@@ -89,7 +93,6 @@ class TaskController extends AbstractController
     public function toggle(Task $task, EntityManagerInterface $em)
     {
         $task->toggle(!$task->isDone());
-//        $this->getDoctrine()->getManager()->flush();
         $em->persist($task);
         $em->flush();
 
@@ -106,10 +109,17 @@ class TaskController extends AbstractController
     #[Route('/tasks/{id}/delete', name: 'task_delete')]
     public function delete(Task $task, EntityManagerInterface $em)
     {
-        $em->remove($task);
-        $em->flush();
+        try {
+            $this->denyAccessUnlessGranted('deleteTask', $task);
 
-        $this->addFlash('success', 'La tâche a bien été supprimée.');
+            $em->remove($task);
+            $em->flush();
+
+            $this->addFlash('success', 'La tâche a bien été supprimée.');
+        } catch (AccessDeniedException) {
+            $this->addFlash('error', 'Accès refusé');
+
+        }
 
         return $this->redirectToRoute('task_list');
     }
